advance notification types
